+++ YAP/YAP/Source/Data/Models/Meal.swift
@@ -9,35 +9,12 @@ import Foundation
 import SwiftData
 
 @Model
-class Menu {
-  var name: String
-  var carbohydrates: Double // 탄수화물
-  var protein: Double // 단백질
-  var lipid: Double // 지방
-  var kcal: Int
-  
-  init(
-    name: String,
-    carbohydrates: Double,
-    protein: Double,
-    lipid: Double,
-    kcal: Int
-  ) {
-    self.name = name
-    self.carbohydrates = carbohydrates
-    self.protein = protein
-    self.lipid = lipid
-    self.kcal = kcal
-  }
-}
-
-@Model
 class Meal {
   var day: Date // 날짜
   var carbohydrates: Double // 탄수화물
   var protein: Double // 단백질
   var lipid: Double // 지방
-  var kcal: Int //칼로리
+  var kcal: Int // 칼로리
   
   @Relationship(deleteRule: .cascade)
   var menus: [Menu]
